Add test for info on order show page
