+++ spec/features/admins/admin_can_see_orders_on_dashboard_spec.rb
@@ -19,25 +19,6 @@ describe "admin sees list of orders on dashboard" do
     expect(page).to have_link "Order #{order2.id} placed #{order2.created_at.to_date}", href: order_path(order2)
   end
 
-  scenario "admin can see other users' order show pages" do
-    user = create(:user)
-    status = Status.create(name: "Completed")
-    order = user.orders.create(status_id: status.id)
-    item = create(:item)
-    order.items << item
-
-    admin = create(:admin)
-
-    allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(admin)
-
-    visit order_path(order)
-
-    expect(page).to have_content(user.name.capitalize)
-    expect(page).to have_content(order.id)
-    expect(page).to have_content(item.title)
-    expect(page).to have_content("Completed")
-  end
-
   scenario "admin sees list of statuses and count on dashboard" do
     admin = create(:admin)
     allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(admin)
